% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{coeff}
\alias{coeff}
\title{coeff}
\usage{
coeff(xk, xk2, data)
}
\arguments{
\item{xk}{the first x knot}

\item{xk2}{the second k knot}

\item{data}{the data for the linear model}
}
\value{
coefficients of the linear model
}
\description{
Creates a linear model using the x knots and data. The function then returns the coefficients of the linear model
}
\examples{
coeff(xk,xk2,data)
}
