% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_rsquared.R
\name{matrix_rsquared}
\alias{matrix_rsquared}
\title{matrix_rsquared}
\usage{
matrix_rsquared(x1, x2, h, data)
}
\arguments{
\item{x1}{the interval for x knot 1}

\item{x2}{the interval for x knot 2}

\item{h}{the size of the interval of each grid}

\item{data}{the data of the linear model
The function that finds the max r^2 withing the confidence interval using grid approximation}
}
\value{
returns x knot 1, x knot 2, and r^2
}
\description{
matrix_rsquared
}
\examples{
matrix_rsquared(x1, x2, 0.001, spruce.df)
}
