% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{A start for a function}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
a vector whose components are squared
}
\description{
Starter function for MATH4773CLAG
}
\examples{
myf(1:10)
}
