% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf2.R
\name{myf2}
\alias{myf2}
\title{myf2}
\usage{
myf2(x, xk, xk2, coef)
}
\arguments{
\item{x}{points on the curve}

\item{xk}{the first x knot}

\item{xk2}{the second k knot}

\item{coef}{linear model coefficients}
}
\value{
coefficients of the linear model used for plotting linear equations
}
\description{
Takes in data and x knot values to create a linear equation.
}
\examples{
myf2(x,xk=input$xk1,xk2 = input$xk2, coef=coef(lmp)
}
